requirements.txt
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cycler==0.11.0
fonttools==4.33.3
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.23.0
odak==0.2.0
packaging==21.3
Pillow==9.1.1
plyfile==0.7.4
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
torch==1.11.0
tqdm==4.64.0
typing_extensions==4.2.0
